move utility usage from config class to s3a util class
